requirements.txt
Django==1.7.1
django-geojson==2.6.0
django-leaflet==0.15.1
django-tastypie==0.12.1
psycopg2==2.5.4
python-dateutil==2.3
python-mimeparse==0.1.4
six==1.8.0
